re-structure src code
