--- src/layouts/post/components/Invoice/index.js
@@ -1,73 +1,0 @@
-/**
-=========================================================
-* Soft UI Dashboard React - v4.0.0
-=========================================================
-
-* Product Page: https://www.creative-tim.com/product/soft-ui-dashboard-react
-* Copyright 2022 Creative Tim (https://www.creative-tim.com)
-
-Coded by www.creative-tim.com
-
- =========================================================
-
-* The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
-*/
-
-// prop-types is a library for typechecking of props
-import PropTypes from "prop-types";
-
-// @mui material components
-import Icon from "@mui/material/Icon";
-
-// Soft UI Dashboard React components
-import SoftBox from "components/SoftBox";
-import SoftTypography from "components/SoftTypography";
-
-function Invoice({ date, id, price, noGutter }) {
-  return (
-    <SoftBox
-      component="li"
-      display="flex"
-      justifyContent="space-between"
-      alignItems="center"
-      py={1}
-      pr={1}
-      mb={noGutter ? 0 : 1}
-    >
-      <SoftBox lineHeight={1}>
-        <SoftTypography display="block" variant="button" fontWeight="medium">
-          {date}
-        </SoftTypography>
-        <SoftTypography variant="caption" fontWeight="regular" color="text">
-          {id}
-        </SoftTypography>
-      </SoftBox>
-      <SoftBox display="flex" alignItems="center">
-        <SoftTypography variant="button" fontWeight="regular" color="text">
-          {price}
-        </SoftTypography>
-        <SoftBox display="flex" alignItems="center" lineHeight={0} ml={3} sx={{ cursor: "poiner" }}>
-          <Icon fontSize="small">picture_as_pdf</Icon>
-          <SoftTypography variant="button" fontWeight="bold">
-            &nbsp;PDF
-          </SoftTypography>
-        </SoftBox>
-      </SoftBox>
-    </SoftBox>
-  );
-}
-
-// Setting default values for the props of Invoice
-Invoice.defaultProps = {
-  noGutter: false,
-};
-
-// Typechecking props for the Invoice
-Invoice.propTypes = {
-  date: PropTypes.string.isRequired,
-  id: PropTypes.string.isRequired,
-  price: PropTypes.string.isRequired,
-  noGutter: PropTypes.bool,
-};
-
-export default Invoice;
