+++ src/mocules/index.js
@@ -1,0 +1,24 @@
+// users
+export {default as UsersOverviewData} from './users/users-overview-data'
+export {default as UsersReportedData} from './users/users-reported-data'
+export {default as UsersBlockedData} from './users/users-blocked-data' 
+
+// post
+export {default as PostPreview} from './post-overview/post-preview'
+export {default as PostList} from './post-overview/post-list'
+export {default as PostDetail} from './post-overview/post-detail'
+export {default as PostComment} from './post-overview/post-comment'
+
+// food
+export {default as FoodPreview} from './food-overview/food-preview'
+export {default as FoodList} from './food-overview/food-list'
+export {default as FoodDetail} from './food-overview/food-detail'
+export {default as FoodComment} from './food-overview/food-comment'
+export {default as FoodContent} from './food-overview/food-content'
+export {default as FoodEvaluation} from './food-overview/food-evaluation'
+
+
+// feedback
+export {default as FeedbackPreview} from './feedback/feedback-preview'
+export {default as FeedbackList} from './feedback/feedback-list'
+
