+++ src/pages/index.js
@@ -1,0 +1,22 @@
+// authen
+export {default as SignInPage} from '../pages/sign-in-page'
+// users
+export {default as UsersOverviewPage} from '../pages/users-overview-page'
+export {default as UsersReportedPage} from '../pages/users-overview-page'
+export {default as UsersBlockedPage} from '../pages/users-overview-page'
+
+// post
+export {default as PostOverviewPage} from '../pages/post-overview-page'
+export {default as PostReportedPage} from '../pages/post-reported-page'
+export {default as PostBlockedPage} from '../pages/post-blocked-page'
+
+// food
+export {default as FoodOverviewPage} from '../pages/food-overview-page'
+export {default as FoodReportedPage} from '../pages/food-overview-page'
+export {default as FoodBlockedPage} from '../pages/food-overview-page'
+
+// setting
+export {default as SettingsPage} from '../pages/settings-page'
+
+// feedback
+export {default as FeedbackPage} from '../pages/feedback-page'
