+++ src/pages/settings-page.js
@@ -1,19 +1,19 @@
 // Soft UI Dashboard React components
 import SoftBox from "components/SoftBox";
+
 import DashboardLayout from "examples/LayoutContainers/DashboardLayout";
 import DashboardNavbar from "examples/Navbars/DashboardNavbar";
 import Footer from "examples/Footer";
 
-function Chat() {
+function SettingsPage() {
     return (
         <DashboardLayout>
-            <DashboardNavbar action='chats' />
+            <DashboardNavbar action='settings' />
             <SoftBox mt={4}>
-
             </SoftBox>
             <Footer />
         </DashboardLayout>
     );
 }
 
-export default Chat;
+export default SettingsPage;
