+++ src/pages/sign-in-page.js
@@ -1,0 +1,98 @@
+import { useState } from "react";
+
+// react-router-dom components
+import { Link } from "react-router-dom";
+
+// @mui material components
+import Switch from "@mui/material/Switch";
+
+// Soft UI Dashboard React components
+import SoftBox from "components/SoftBox";
+import SoftTypography from "components/SoftTypography";
+import SoftInput from "components/SoftInput";
+import SoftButton from "components/SoftButton";
+import API from "services/axiosClient"
+
+// Authentication layout components
+import CoverLayout from "layouts/authentication/components/CoverLayout";
+
+// Images
+import curved9 from "assets/images/curved-images/curved-6.jpg";
+
+function SignInPage() {
+  const [rememberMe, setRememberMe] = useState(true);
+
+  const handleSetRememberMe = () => setRememberMe(!rememberMe);
+
+  const _handleLogin = async () => {
+    try {
+      const data = await API.login({
+        email: "nguyennhuttan12201@gmail.com",
+        password: "123456789"
+      })
+      console.log("Data: ", data)
+    } catch (error) {
+      console.log("Error: ", error)
+    }
+  }
+
+  return (
+    <CoverLayout
+      top={8}
+      bottom={8}
+      title="Welcome back"
+      description="Enter your email and password to sign in with Foodtalk"
+    >
+      <SoftBox component="form" role="form">
+        <SoftBox mb={2}>
+          <SoftBox mb={1} ml={0.5}>
+            <SoftTypography component="label" variant="caption" fontWeight="bold">
+              Email
+            </SoftTypography>
+          </SoftBox>
+          <SoftInput type="email" placeholder="Email" />
+        </SoftBox>
+        <SoftBox mb={2}>
+          <SoftBox mb={1} ml={0.5}>
+            <SoftTypography component="label" variant="caption" fontWeight="bold">
+              Password
+            </SoftTypography>
+          </SoftBox>
+          <SoftInput type="password" placeholder="Password" />
+        </SoftBox>
+        <SoftBox display="flex" alignItems="center">
+          <Switch checked={rememberMe} onChange={handleSetRememberMe} />
+          <SoftTypography
+            variant="button"
+            fontWeight="regular"
+            onClick={handleSetRememberMe}
+            sx={{ cursor: "pointer", userSelect: "none" }}
+          >
+            &nbsp;&nbsp;Show password
+          </SoftTypography>
+        </SoftBox>
+        <SoftBox mt={4} mb={1}>
+          <SoftButton onClick={_handleLogin} variant="gradient" color="primary" fullWidth>
+            sign in
+          </SoftButton>
+        </SoftBox>
+        <SoftBox mt={3} textAlign="center">
+          <SoftTypography variant="button" color="text" fontWeight="regular">
+            Don&apos;t have an account?{" "}
+            <SoftTypography
+              component={Link}
+              to="/authentication/sign-up"
+              variant="button"
+              color="primary"
+              fontWeight="medium"
+            >
+              Sign up
+            </SoftTypography>
+          </SoftTypography>
+        </SoftBox>
+      </SoftBox>
+    </CoverLayout>
+  );
+}
+
+export default SignInPage;
