+++ src/pages/users-overview-page.js
@@ -1,0 +1,70 @@
+/**
+=========================================================
+* Soft UI Dashboard React - v4.0.0
+=========================================================
+
+* Product Page: https://www.creative-tim.com/product/soft-ui-dashboard-react
+* Copyright 2022 Creative Tim (https://www.creative-tim.com)
+
+Coded by www.creative-tim.com
+
+ =========================================================
+
+* The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+*/
+
+// @mui material components
+import Card from "@mui/material/Card";
+
+// Soft UI Dashboard React components
+import SoftBox from "components/SoftBox";
+import SoftTypography from "components/SoftTypography";
+
+// Soft UI Dashboard React examples
+import DashboardLayout from "examples/LayoutContainers/DashboardLayout";
+import DashboardNavbar from "examples/Navbars/DashboardNavbar";
+import Footer from "examples/Footer";
+import Table from "examples/Tables/Table";
+import { useEffect } from "react";
+import { RootStore } from "store/RootStore";
+import { observer } from "mobx-react-lite";
+import { UsersOverviewData } from "mocules";
+
+function UsersOverviewPage() {
+  const { getUsers, users } = RootStore;
+  const { columns, rows} = UsersOverviewData(users);
+
+  useEffect(() => {
+    getUsers();
+  }, []);
+
+  return (
+    <DashboardLayout>
+      <DashboardNavbar action="users" />
+      <SoftBox py={3}>
+        <SoftBox mb={3}>
+          <Card>
+            <SoftBox display="flex" justifyContent="space-between" alignItems="center" p={3}>
+              <SoftTypography variant="h6">Users</SoftTypography>
+            </SoftBox>
+            <SoftBox
+              sx={{
+                "& .MuiTableRow-root:not(:last-child)": {
+                  "& td": {
+                    borderBottom: ({ borders: { borderWidth, borderColor } }) =>
+                      `${borderWidth[1]} solid ${borderColor}`,
+                  },
+                },
+              }}
+            >
+              <Table columns={columns} rows={rows} />
+            </SoftBox>
+          </Card>
+        </SoftBox>
+      </SoftBox>
+      <Footer />
+    </DashboardLayout>
+  );
+}
+
+export default observer(UsersOverviewPage);
