+++ src/pages/users-reported-page.js
@@ -1,0 +1,30 @@
+/**
+=========================================================
+* Soft UI Dashboard React - v4.0.0
+=========================================================
+
+* Product Page: https://www.creative-tim.com/product/soft-ui-dashboard-react
+* Copyright 2022 Creative Tim (https://www.creative-tim.com)
+
+Coded by www.creative-tim.com
+
+ =========================================================
+
+* The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+*/
+// Soft UI Dashboard React examples
+import DashboardLayout from "examples/LayoutContainers/DashboardLayout";
+import DashboardNavbar from "examples/Navbars/DashboardNavbar";
+import { observer } from "mobx-react-lite";
+
+function UsersReportedPage() {
+
+  return (
+    <DashboardLayout>
+      <DashboardNavbar action="users" />
+      
+    </DashboardLayout>
+  );
+}
+
+export default observer(UsersReportedPage);
